FeedParser's ``_LooseFeedParser`` class is (seemingly in pretty much all cases) instantiated with an empty string as the ```encoding`` value, which can lead to exceptions down the road when the encoding is actually used.
Ironically, the loose parser seems to be used primarily for cases where encoding problems (like not being able to find a valid, working encoding) make using the strict parser impossible in the first place.

This is reported as issue #51, and includes a patch by blaker, on which our own patch is based. See this ticket for a whole bunch of additional information.

While not strictly necessary to run FeedPlatform, you are encouraged to use the patch. Occurance of this bug is somewhat rare, but can easily happen.