Make sure that everything is exposed as unicode strings, even when the feed is bozo.
Fixes a bug that can cause certain values, mostly when taken from element attributes, to be byte strings when the feed is bozo.

The problem is that unlike tag contents, which are already normalized to unicode, certain vales are taken directly, as-is, from whatever is returned by the parser. Depending on what parser is used (loose, strict, ...), the behaviour can differ, and the loose parser seems to return bytestrings.

Affected are for example:

  - enclosure data
  - feed image url when taken from itunes:image

While the patch is not strictly required, you are strongly encouraged to use it. FeedPlatform expects the parser to expose all values as unicode, and it's ORM (Storm) requires unicode input. Without the patch, you are bound to see ORM exceptions sooner or later.

Reported as issue #148.