No longer convert an item id (guid) into an absolute link if isPermaLink=="false".
While not strictly necessary to run FeedPlatform, it is strongly recommended. For example, one significant issue arising from this is that the domain name used to build the absolute url may change (say the feed is redirected to a new address), thus effectively changing all the ids.

Reported as issue #107.