FeedParser, in trunk, breaks when an item (invalidly) contains an <image> tag.
The 4.1 release did not exhibit this bug.

This is reported as issue #108, and includes a patch by jabronson, on which our own patch is based.

While not strictly necessary to run FeedPlatform, you are strongly encouraged to use the patch, or you'll probably run into the bug sooner or later.